Moved version.py into the hatchet directory.
